restructuring runscript, fixing # 322
