+++ avaframe/in3Utils/initializeProject.py
@@ -20,7 +20,7 @@ def _checkForFolderAndDelete(baseDir, folderName):
         log.debug("No %s folder found.", folderName)
 
 
-def cleanSingleAvaDir(avaDir, keep=None, deleteOutput=True):
+def cleanSingleAvaDir(avaDir, keep=None, deleteOutput=True, onlyModule=''):
     '''
     Clean a single avalanche directory of the work and output directories
     Expects a avalanche directory name as string
@@ -46,7 +46,11 @@ def cleanSingleAvaDir(avaDir, keep=None, deleteOutput=True):
     # Try to remove OUTPUTS folder, only pass FileNotFoundError, i.e. folder
     # does not exist
     if deleteOutput:
-        _checkForFolderAndDelete(avaDir, 'Outputs')
+        if onlyModule:
+            fPath = os.path.join(avaDir, 'Outputs')
+            _checkForFolderAndDelete(fPath, onlyModule)
+        else:
+            _checkForFolderAndDelete(avaDir, 'Outputs')
 
     # Try to remove Work folder, only pass FileNotFoundError, i.e. folder
     # does not exist
